fix(teahouse): Vault token query filter optimization
